Extract argument processing code from c_main macro
Since Rust macros are expanded before any compilation takes place, having code that actually does stuff in the macro increases compile time for group compiles. This can easily be solved by moving the code outside the macro into a function that is shared by each of the binaries.

I simply extracted the code into a function and referenced that function inside the macro.